requirements.txt
# AUVAP Framework Requirements

# CyberBattleSim - Base Environment
# Note: CyberBattleSim must be installed separately from GitHub:
# pip install git+https://github.com/microsoft/CyberBattleSim.git
gym>=0.21.0
networkx>=2.6.3

# Reinforcement Learning
stable-baselines3>=2.0.0
torch>=2.0.0
tensorboard>=2.12.0

# Graph Database
neo4j>=5.12.0
py2neo>=2021.2.3

# Data Processing & Visualization
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
pyyaml>=6.0
python-dotenv>=1.0.0
loguru>=0.7.0
tqdm>=4.65.0

# Testing & Development
pytest>=7.3.0
pytest-cov>=4.1.0
black>=23.3.0
flake8>=6.0.0
